refactor: split out class field test into more scoped test files
