+++ tools/serverpod_cli/test/generator/dart/server_code_generator/entity_class_test.dart
@@ -100,30 +100,6 @@ void main() {
     }, skip: maybeClassNamedExample == null);
   });
 
-  group('Given class with documentation when generating code', () {
-    var documentation = [
-      '// This is an example documentation',
-      '// This is another example'
-    ];
-    var entities = [
-      ClassDefinitionBuilder()
-          .withFileName(testClassFileName)
-          .withDocumentation(documentation)
-          .build()
-    ];
-
-    var codeMap = generator.generateSerializableEntitiesCode(
-      entities: entities,
-      config: config,
-    );
-
-    test('then documentation is included in code.', () {
-      for (var comment in documentation) {
-        expect(codeMap[expectedFilePath], contains(comment));
-      }
-    });
-  });
-
   group('Given a class with table name when generating code', () {
     var entities = [
       ClassDefinitionBuilder()
@@ -332,339 +308,6 @@ void main() {
             reason: 'Missing declaration for static include method.');
       });
     }, skip: maybeClassNamedExample == null);
-
-    var maybeClassNamedExampleTable =
-        CompilationUnitHelpers.tryFindClassDeclaration(
-      compilationUnit,
-      name: '${testClassName}Table',
-    );
-
-    test('then a class named ${testClassName}Table is generated.', () {
-      expect(
-        maybeClassNamedExampleTable,
-        isNotNull,
-        reason: 'Missing definition for class named ${testClassName}Table',
-      );
-    });
-    group('then the class named ${testClassName}Table', () {
-      test('inherits from Table.', () {
-        expect(
-            CompilationUnitHelpers.hasExtendsClause(
-              maybeClassNamedExampleTable!,
-              name: 'Table',
-            ),
-            isTrue,
-            reason: 'Missing extends clause for Table.');
-      });
-
-      test(
-          'has constructor taking query prefix and table relations and passes table name to super.',
-          () {
-        expect(
-            CompilationUnitHelpers.hasConstructorDeclaration(
-              maybeClassNamedExampleTable!,
-              name: null,
-              parameters: ['super.queryPrefix', 'super.tableRelations'],
-              superArguments: ['tableName: \'$tableName\''],
-            ),
-            isTrue,
-            reason:
-                'Missing declaration for $testClassName constructor with nullable id field passed to super.');
-      });
-
-      test('has a columns method.', () {
-        expect(
-            CompilationUnitHelpers.hasMethodDeclaration(
-              maybeClassNamedExampleTable!,
-              name: 'columns',
-            ),
-            isTrue,
-            reason: 'Missing declaration for columns getter.');
-      });
-
-      test('does NOT have getRelationTable method.', () {
-        expect(
-            CompilationUnitHelpers.hasMethodDeclaration(
-              maybeClassNamedExampleTable!,
-              name: 'getRelationTable',
-            ),
-            isFalse,
-            reason:
-                'Declaration for getRelationTable method should not be generated.');
-      });
-
-      test('does NOT have id field.', () {
-        expect(
-            CompilationUnitHelpers.hasFieldDeclaration(
-              maybeClassNamedExampleTable!,
-              name: 'id',
-            ),
-            isFalse,
-            reason: 'Declaration for id field should not be generated.');
-      });
-    }, skip: maybeClassNamedExampleTable == null);
-
-    test(
-        'then top level variable t$testClassName marked deprecated is generated.',
-        () {
-      expect(
-          CompilationUnitHelpers.hasTopLevelVariableDeclaration(
-            compilationUnit,
-            name: 't$testClassName',
-            annotations: ['Deprecated'],
-          ),
-          isTrue,
-          reason:
-              'Missing top level variable declaration for "t$testClassName" marked deprecated.');
-    });
-
-    test('then type alias ${testClassName}ExpressionBuilder is generated.', () {
-      expect(
-          CompilationUnitHelpers.hasTypeAliasDeclaration(
-            compilationUnit,
-            name: '${testClassName}ExpressionBuilder',
-          ),
-          isTrue,
-          reason:
-              'Missing type alias for "${testClassName}ExpressionBuilder".');
-    });
-
-    test('then a class named ${testClassName}Include is generated.', () {
-      expect(
-          CompilationUnitHelpers.hasClassDeclaration(
-            compilationUnit,
-            name: '${testClassName}Include',
-          ),
-          isTrue,
-          reason: 'Missing class named ${testClassName}Include.');
-    });
-  });
-
-  group(
-      'Given a class with table name and persistent field when generating code',
-      () {
-    var entities = [
-      ClassDefinitionBuilder()
-          .withClassName(testClassName)
-          .withFileName(testClassFileName)
-          .withTableName(tableName)
-          .withField(
-            FieldDefinitionBuilder()
-                .withName('title')
-                .withTypeDefinition('String', true)
-                .withScope(EntityFieldScopeDefinition.all)
-                .withShouldPersist(true)
-                .build(),
-          )
-          .build()
-    ];
-
-    var codeMap = generator.generateSerializableEntitiesCode(
-      entities: entities,
-      config: config,
-    );
-
-    var compilationUnit = parseString(content: codeMap[expectedFilePath]!).unit;
-    var maybeClassNamedExampleTable =
-        CompilationUnitHelpers.tryFindClassDeclaration(
-      compilationUnit,
-      name: '${testClassName}Table',
-    );
-
-    group('then the class named ${testClassName}Table', () {
-      test('has class variable for field.', () {
-        expect(
-            CompilationUnitHelpers.hasFieldDeclaration(
-              maybeClassNamedExampleTable!,
-              name: 'title',
-              isFinal: true,
-              isLate: true,
-            ),
-            isTrue,
-            reason: 'Missing declaration for title field.');
-      });
-
-      test('has field included in columns.', () {
-        expect(
-            CompilationUnitHelpers.hasMethodDeclaration(
-              maybeClassNamedExampleTable!,
-              name: 'columns',
-              functionExpression: '[id, title]',
-            ),
-            isTrue,
-            reason: 'Missing title field in columns.');
-      });
-    },
-        skip: maybeClassNamedExampleTable == null
-            ? 'Could not run test because ${testClassName}Table class was not found'
-            : false);
-  });
-
-  group(
-      'Given a class with table name and NON persistent field when generating code',
-      () {
-    var entities = [
-      ClassDefinitionBuilder()
-          .withClassName(testClassName)
-          .withFileName(testClassFileName)
-          .withTableName(tableName)
-          .withField(
-            FieldDefinitionBuilder()
-                .withName('title')
-                .withTypeDefinition('String', true)
-                .withScope(EntityFieldScopeDefinition.all)
-                .withShouldPersist(false)
-                .build(),
-          )
-          .build()
-    ];
-
-    var codeMap = generator.generateSerializableEntitiesCode(
-      entities: entities,
-      config: config,
-    );
-
-    var compilationUnit = parseString(content: codeMap[expectedFilePath]!).unit;
-    var maybeClassNamedExampleTable =
-        CompilationUnitHelpers.tryFindClassDeclaration(
-      compilationUnit,
-      name: '${testClassName}Table',
-    );
-
-    group('then the class named ${testClassName}Table', () {
-      test(
-        'does NOT have class variable for field.',
-        () {
-          expect(
-              CompilationUnitHelpers.hasFieldDeclaration(
-                maybeClassNamedExampleTable!,
-                name: 'title',
-                isFinal: true,
-              ),
-              isFalse,
-              reason: 'Should not have declaration for title field.');
-        },
-      );
-
-      test('does NOT have field included in columns.', () {
-        expect(
-            CompilationUnitHelpers.hasMethodDeclaration(
-              maybeClassNamedExampleTable!,
-              name: 'columns',
-              functionExpression: '[id]',
-            ),
-            isTrue,
-            reason: 'Should not include field in columns.');
-      });
-    },
-        skip: maybeClassNamedExampleTable == null
-            ? 'Could not run test because ${testClassName}Table class was not found.'
-            : false);
-  });
-
-  group(
-      'Given a class with table name and object relation field when generating code',
-      () {
-    var entities = [
-      ClassDefinitionBuilder()
-          .withClassName(testClassName)
-          .withFileName(testClassFileName)
-          .withTableName(tableName)
-          .withObjectRelationField('company', 'Company', 'company')
-          .build()
-    ];
-
-    var codeMap = generator.generateSerializableEntitiesCode(
-      entities: entities,
-      config: config,
-    );
-
-    var compilationUnit = parseString(content: codeMap[expectedFilePath]!).unit;
-    var maybeClassNamedExampleInclude =
-        CompilationUnitHelpers.tryFindClassDeclaration(
-      compilationUnit,
-      name: '${testClassName}Include',
-    );
-
-    group('then the class named ${testClassName}Include', () {
-      var exampleIncludeClass = maybeClassNamedExampleInclude!;
-      test('inherits from Include.', () {
-        expect(
-            CompilationUnitHelpers.hasExtendsClause(
-              exampleIncludeClass,
-              name: 'Include',
-            ),
-            isTrue,
-            reason: 'Missing extends clause for Include.');
-      });
-      test('has named parameter for field in private constructor.', () {
-        expect(
-            CompilationUnitHelpers.hasConstructorDeclaration(
-              exampleIncludeClass,
-              name: '_',
-              parameters: ['CompanyInclude? company'],
-            ),
-            isTrue,
-            reason:
-                'Missing constructor with named parameter for field in ${testClassName}Include.');
-      });
-
-      test('has private field as nullable class variable.', () {
-        expect(
-            CompilationUnitHelpers.hasFieldDeclaration(
-              exampleIncludeClass,
-              name: '_company',
-              type: 'CompanyInclude?',
-            ),
-            isTrue,
-            reason:
-                'Missing declaration for company field in ${testClassName}Include.');
-      });
-      test('has an includes method.', () {
-        expect(
-            CompilationUnitHelpers.hasMethodDeclaration(
-              exampleIncludeClass,
-              name: 'includes',
-            ),
-            isTrue,
-            reason: 'Missing declaration for includes method.');
-      });
-
-      test('has a table method.', () {
-        expect(
-            CompilationUnitHelpers.hasMethodDeclaration(
-              exampleIncludeClass,
-              name: 'table',
-              functionExpression: 'Example.t',
-            ),
-            isTrue,
-            reason: 'Missing declaration for table method.');
-      });
-    },
-        skip: maybeClassNamedExampleInclude == null
-            ? 'Could not run test because ${testClassName}Include class was not found.'
-            : false);
-
-    var maybeClassNamedExampleTable =
-        CompilationUnitHelpers.tryFindClassDeclaration(
-      compilationUnit,
-      name: '${testClassName}Table',
-    );
-
-    test(
-        'then the class named ${testClassName}Table has a getRelationTable method.',
-        () {
-      expect(
-          CompilationUnitHelpers.hasMethodDeclaration(
-            maybeClassNamedExampleTable!,
-            name: 'getRelationTable',
-          ),
-          isTrue,
-          reason: 'Missing declaration for getRelationTable method.');
-    },
-        skip: maybeClassNamedExampleTable == null
-            ? 'Could not run test because ${testClassName}Table class was not found.'
-            : false);
   });
 
   group('Given a class with a none nullable field when generating code', () {
